+++ .gitignore
@@ -1,7 +1,5 @@
 # TODO: cleanup
 php-error.log
-build.sh
-debug.sh
 run.sh
 test.sh
 valgrind.sh
