+++ firebird_utils.h
@@ -1,19 +1,3 @@
-/*
-  +----------------------------------------------------------------------+
-  | Copyright (c) The PHP Group                                          |
-  +----------------------------------------------------------------------+
-  | This source file is subject to version 3.01 of the PHP license,      |
-  | that is bundled with this package in the file LICENSE, and is        |
-  | available through the world-wide-web at the following url:           |
-  | https://www.php.net/license/3_01.txt                                 |
-  | If you did not receive a copy of the PHP license and are unable to   |
-  | obtain it through the world-wide-web, please send a note to          |
-  | license@php.net so we can mail you a copy immediately.               |
-  +----------------------------------------------------------------------+
-  | Author: Simonov Denis <sim-mail@list.ru>                             |
-  +----------------------------------------------------------------------+
-*/
-
 #ifndef FIREBIRD_UTILS_H
 #define FIREBIRD_UTILS_H
 
@@ -23,7 +7,7 @@ extern "C" {
 
 #include <ibase.h>
 #include "database.h"
-#include "fbp_database.h"
+#include "database.h"
 #include "transaction.h"
 #include "fbp_statement.h"
 #include "fbp_blob.h"
